src/requirements.txt
flask>=3.0
gunicorn>=22.0
azure-ai-projects>=2.0.0b1
azure-identity>=1.17.0
